+++ NEWS.md
@@ -1,4 +1,5 @@
 # Changements v 0.1.1
 
 * Ajout de fonction pour télécharger les données de hotspots eBird
+* Ajout de description pour la formation de la grille.